feat(frontend): refactor handleSideEffect into an util file
